lib: video: Create local abstraction
